* subversion/libsvn_client/diff.c   (transform_link_to_git): Use safer type for formatting.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1706625 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/diff.c
@@ -762,7 +762,7 @@ transform_link_to_git(const char **new_tmpfile,
 
       /* git calculates the sha over "blob X\0" + the actual data,
          where X is the decimal size of the blob. */
-      sz_str = apr_psprintf(scratch_pool, "blob %u", line->len);
+      sz_str = apr_psprintf(scratch_pool, "blob %u", (unsigned int)line->len);
       svn_stringbuf_insert(line, 0, sz_str, strlen(sz_str) + 1);
 
       SVN_ERR(svn_checksum(&checksum, svn_checksum_sha1,
